requirements.txt
opencv-python
cvzone
numpy
keras
tensorflow
gtts
pyenchant